AI mini projects/Lenet/requirements.txt
numpy
matplotlib
opencv-python
torch
torchvision
easydict
python-dotenv
azureml-defaults
azureml-core
azureml-train
mlflow
azureml-mlflow
azureml-pipeline
azureml-widgets
azureml-contrib-services
